+++ test/scripts.spec.js
@@ -4,7 +4,25 @@ import {describe, it} from 'node:test'
 import {runViteBuild} from './run.js'
 
 describe('<script vite-inline>', () => {
-    it('inlines javascript', async () => {
+    it('[vite-inline="-minify"] inlines javascript  as is without esbuild', async () => {
+        for (const src of ['/index.js', './index.js', 'index.js']) {
+            const result = await runViteBuild({
+                files: {
+                    'index.html': `<html><script vite-inline="-minify" src="${src}"></script></html>`,
+                    'index.js': `console.log(location.search)`,
+                },
+            })
+            assert.ok(result.success, result.output)
+            assert.equal(
+                await result.fromDistContent('index.html'),
+                '<html><script>console.log(location.search)</script></html>',
+                result.root,
+            )
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
+            await rm(result.root, {recursive: true})
+        }
+    })
+    it('inlines javascript with default esbuild minify', async () => {
         for (const src of ['/index.js', './index.js', 'index.js']) {
             const result = await runViteBuild({
                 files: {
@@ -15,15 +33,15 @@ describe('<script vite-inline>', () => {
             assert.ok(result.success, result.output)
             assert.equal(
                 await result.fromDistContent('index.html'),
-                '<html><script>console.log(location.search)</script></html>',
+                '<html><script>console.log(location.search);</script></html>',
                 result.root,
             )
-            assert.deepStrictEqual(['index.html'], await result.distFiles(), result.root)
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
             await rm(result.root, {recursive: true})
         }
     })
 
-    it('inlines javascript type="module"', async () => {
+    it('inlines javascript type="module" with esbuild minify', async () => {
         for (const src of ['/index.js', './index.js', 'index.js']) {
             const result = await runViteBuild({
                 files: {
@@ -33,14 +51,9 @@ describe('<script vite-inline>', () => {
                 },
             })
             assert.ok(result.success, result.output)
-            assert.equal(
-                await result.fromDistContent('index.html'),
-                `<html><script  type="module">// delegate.js
-console.log(location.search);
-</script></html>`,
-                result.root,
-            )
-            assert.deepStrictEqual(['index.html'], await result.distFiles(), result.root)
+            assert.ok(await result.fromDistTest('index.html', /<script>[.\s\S]+<\/script>/), result.root)
+            assert.ok(await result.fromDistOmits('index.html', 'import'), result.root)
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
             await rm(result.root, {recursive: true})
         }
     })
@@ -64,22 +77,9 @@ console.log(location.search);
                 },
             })
             assert.ok(result.success, result.output)
-            assert.equal(
-                await result.fromDistContent('index.html'),
-                `<html><script>// index.ts
-var Foo = class {
-  constructor(asdf) {
-    this.asdf = asdf;
-  }
-  isAsdfy() {
-    return !!this.asdf;
-  }
-};
-new Foo("0").isAsdfy();
-</script></html>`,
-                result.root,
-            )
-            assert.deepStrictEqual(['index.html'], await result.distFiles(), result.root)
+            assert.ok(await result.fromDistTest('index.html', /<script>[.\s\S]+<\/script>/), result.root)
+            assert.ok(await result.fromDistOmits('index.html', 'Foo'), result.root)
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
             await rm(result.root, {recursive: true})
         }
     })
@@ -103,14 +103,30 @@ new Foo("0").isAsdfy();
                     `,
                 },
             })
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
             assert.ok(result.success, result.output)
-            assert.ok(await result.fromDistOmits('index.html', 'getDeviceVersion'), result.root)
-            assert.deepStrictEqual(['index.html'], await result.distFiles(), result.root)
+            assert.ok(await result.fromDistOmits('index.html', '8675309'), result.root)
+            assert.deepEqual(['index.html'], await result.distFiles(), result.root)
             await rm(result.root, {recursive: true})
         }
     })
 
     describe('errors', () => {
+        it('for unknown vite-inline flags', async () => {
+            const result = await runViteBuild({
+                files: {
+                    'index.html': `
+                        <html>
+                            <script vite-inline="jolly roger" src="index.js" fetchpriority="high"></script>
+                        </html>`,
+                    'index.js': `console.log(location.search)`,
+                },
+            })
+            assert.ok(result.error, result.output)
+            assert.ok(result.output.includes('<script vite-inline="jolly"> is an unknown vite-inline flag'), result.output)
+            await rm(result.root, {recursive: true})
+        })
+
         it('for additional attributes', async () => {
             const result = await runViteBuild({
                 files: {
@@ -173,11 +189,23 @@ new Foo("0").isAsdfy();
             assert.ok(result.output.includes('<script vite-inline> must use a relative filesystem src path (network paths like https://cdn/index.js are unsupported)'), result.output)
             await rm(result.root, {recursive: true})
         })
-
-        it('for missing file', async () => {
+        
+        it('with esbuild for missing file', async () => {
             const result = await runViteBuild({
                 files: {
                     'index.html': '<html><script vite-inline src="index.js"></script></html>',
+                },
+            })
+            assert.ok(result.error, result.output)
+            assert.ok(result.output.includes('esbuild processing index.js'), result.output)
+            assert.ok(result.output.includes('Could not resolve'), result.output)
+            await rm(result.root, {recursive: true})
+        })
+
+        it('without esbuild for missing file', async () => {
+            const result = await runViteBuild({
+                files: {
+                    'index.html': '<html><script vite-inline="-minify" src="index.js"></script></html>',
                 },
             })
             assert.ok(result.error, result.output)